fix(core): remove type declaration from ts file
