+++ packages/core/scripts/custom-elements/custom-elements.d.ts
@@ -6,5 +6,5 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-export * from '../dist/types/index';
+export * from '../dist/types/interfaces';
 export * from './index';
