+++ packages/core/src/components/playground/playground.scss
