+++ packages/core/src/index.html
@@ -30,6 +30,6 @@ SPDX-License-Identifier: MIT
 </head>
 
 <body style="margin: 0px; height: 100vh">
-  <my-component />
+  <ix-playground-internal />
 </body>
 </html>
